Fix race conditions in NamedPipe and clean it up.
The named pipe is marked FILE_FLAG_OVERLAPPED, but we're doing I/O without
an OVERLAPPED structure.  MSDN states that this isn't allowed.  This blog
post[1] explains that it works, but only if the application is careful
never to issue two I/O operations at the same time.  The NamedPipe class
tries to achieve mutual exclusions using non-volatile boolean flags and
polling for available data.  In principle, this doesn't work, because AFAIK
Java has weak guarantees about non-volatile variables, and because the
threads could be interrupted immediately prior to the assignments to
readNotify and writeNotify.

With the current code, I can hang NamedPipe easily.  To reproduce, I
run IntelliJ or jediterm in a Windows VM, run "dir /s c:\", then mash the
keyboard for 5-60 seconds.  Eventually, the directory listing halts, and
nothing I do can unfreeze it.  I can observe that keypresses are still
getting through to the winpty-agent (using WINPTYDBG=1 and
winpty-debugserver).  From my debugging, it looks like the Java VM gets
stuck in this loop:

    while (writeNotify) {
    }//wait for write to finish

Nonetheless, I can see that writeNotify is being set to false.  writeNotify
isn't volatile, so I think Java is allowed to factor out the read, turning
the code into this:

    if (writeNotify) {
        while (true) {}
    }

[1] https://blogs.msdn.microsoft.com/oldnewthing/20121012-00/?p=6343

I've rewritten the NamedPipe class to instead have a read event, a write
event, and a shutdown event.  I/O blocking is implemented using
WaitForMultipleObjects() on the I/O event and the shutdown event.  A
concurrent read (or write) waits for existing reads (or writes) to complete
before beginning, but a read and a write may run concurrently.  When the
pipe is closed (or marked closed), the closeImpl() function sets the
shutdown event.  If any I/O operations are pending (a read is likely), then
the thread waiting on the I/O operation calls CancelIo().  Once all I/O
threads have finished, the close() function releases resources.

Aside from fixing the read-write race condition, it also stops available()
from calling PeekNamedPipe() with an invalid HANDLE value (which could have
been reopened in the meantime, in principle).

While writing this code, I noticed an interesting problem with the
ReadFile() and GetOverlappedResult() functions, which were declared to take
a WinBase.OVERLAPPED parameter.  A pending read operation would complete
successfully, but the amount of data read was invalid.  I fixed the problem
by changing the two functions to instead accept a Pointer parameter.  I
think the problem is that a parameter declared as WinBase.OVERLAPPED is
passed *by value*, but we need to pass it by reference, because Windows is
going to modify the structure asynchronously.  Passing WinBase.OVERLAPPED
*almost* works because the calling convention converts it to a pointer.
The address can change between calls, though.

I've tried to preserve the existing error handling behavior, roughly.
Details:

 * read(): Previously, read0() didn't throw IOException, because
   WinPty.KERNEL32.ReadFile() doesn't, and if ReadFile() failed, then
   read0() returned the amount of data transferred (presumably 0).  When
   the pipe was closed during a read() call, read() either returned 0
   (myHandle == null) or returned -1 (available() calls PeekNamedPipe()
   with an invalid HANDLE, throws an exception, read() catches it).  If
   Thread.sleep() were interrupted somehow, then read() returned -1.

   The only caller to read() (in WinPTYInputStream) considered all return
   values <= 0 to mean EOF and returned -1.  Therefore, I've made
   NamedPiped.read() simply return -1 on any kind of error or EOF
   condition.

 * write(): Previously, write() always returned wSuccess, which was always
   false.  write0() didn't throw an IOException, because
   Kernel32.INSTANCE.WriteFile() doesn't.  The only caller to write() (in
   WinPTYOutputStream) ignores the return value.  I changed
   NamedPipe.write() to return void.

 * close() threw an IOException if the pipe handle cannot be closed.  I
   preserved this throw.

Fixes https://github.com/traff/pty4j/issues/25