Update pty4j to work with the latest winpty version
Details:

Remove parts of the WinPTY{Input,Output}Stream code.  Some of the code is
redundant with what NamedPipe already does.

Add a getChildProcessId method to WinPty and WinPtyProcess.

Ensure that the process handle stays open until both of these conditions
are met:
 1. The process has exited.
 2. We have disconnected from the agent, by closing the winpty_t
    object.

For consistency with the getChildProcessId method in pty4j-0.6, and to
reduce the possibility of a race condition, have getChildProcessId return
-1 as soon as the WinPty object is closed, even if the process is still
running (for a brief moment).

Close the WinPty object once all the WinPTYInputStream's are closed

The UnixPtyProcess object has similar logic.  If I understand it correctly,
in UnixPtyProcess, closing either getInputStream() or getOutputStream()
closes the main pty, which then prevents I/O on either of the stream.
Closing getErrorStream() closes the error pty.

In winpty's case, we have one expensive resource (the winpty agent
process), and because it has output queues for both CONOUT and CONERR, we
don't want to kill it until we're done with both CONOUT and CONERR.

In WinPTYOutputStream, keep a reference to WinPty to extend its lifetime.