Fix issue with reading process output of terminated process
This fixes the issue on macOS that was happening if a process was exiting too early, and we were unable to read its output from the master file descriptor. The reason is automatic closing of the slave file descriptor on process termination.

So, we have to keep an open file descriptor pointing to the slave side of the pipe.

See this link for details: https://unix.stackexchange.com/a/538271

The issue has been reported to Apple: https://developer.apple.com/forums/thread/663632