Fix EOF handling in the macOS Intel launcher
When using the new isSpawnProcessUsingJdkOnMacIntel
option, a Reaper thread isn't created. Most of the stuff
it does isn't necessary for JDK launchers, but one thing is:
we still must manually wake up stdout/stderr readers,
otherwise they may end up blocked forever, even after
the process has finished.

Add a test for this. Seems it only happens with a certain
combination of factors which are set up in the test.
Particularly, using the console mode is necessary for
this bug to happen.